Update README.md
The aim and objectives of this application is to use a natural device free interface, which recognizes the hand gestures as commands. The application uses a webcam which is used for image acquisition. To control the VLC media player using a defined gesture, the application focuses on some functions of VLC which are used more frequently